// interfaces/enums.ts
export enum ICriteria {
  GENERAL = "GENERAL",
  CITATIONS = "CITATIONS",
  INTERNATIONAL_OUTLOOK = "INTERNATIONAL_OUTLOOK",
  TEACHING = "TEACHING",
  RESEARCH = "RESEARCH",
  STUDENTS_EVALUATION = "STUDENTS_EVALUATION",
}

export enum StateTypes {
  ABIA = "ABIA",
  ANAMBRA = "ANAMBRA",
  ENUGU = "ENUGU",
  EBONYI = "EBONYI",
  IMO = "IMO",
}

export enum LecturerLevel {
  JUNIOR = "JUNIOR",
  SENIOR = "SENIOR",
}

export enum OwnerTypes {
  STATE = "STATE",
  FEDERAL = "FEDERAL",
  PRIVATE = "PRIVATE",
}

export enum AccountTypes {
  GUEST = "GUEST",
  OWNER = "OWNER",
  ADMIN = "ADMIN",
  LECTURER = "LECTURER",
  STUDENT = "STUDENT",
}

export enum AccountRoles {
  OWNER = "OWNER",
  ADMIN = "ADMIN",
  VC = "VC",
  HOD = "HOD",
  USER = "USER",
}

export enum Gender {
  NOTSAY = "NOTSAY",
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum StudentType {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}

export enum LecturerType {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}